+++ internal/impl/confluent/serde_goavro_test.go
@@ -109,7 +109,7 @@ func TestAvroReferences(t *testing.T) {
 
 			cfg := decodingConfig{}
 			cfg.avro.rawUnions = true
-			decoder, err := newSchemaRegistryDecoder(urlStr, noopReqSign, nil, cfg, service.MockResources())
+			decoder, err := newSchemaRegistryDecoder(urlStr, noopReqSign, nil, cfg, schemaStaleAfter, service.MockResources())
 			require.NoError(t, err)
 
 			t.Cleanup(func() {