+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- New `cache_duration` field to `schema_registry_decode`. (@rockwotj)
 - (Benthos) Field `client_auth` added to the `socket_server` input. (@filippog)
 - (Benthos) New Bloblang string method `uuid_v5`. (@artemklevtsov)
 - New `qdrant` processor. (@rockwotj)